Extensions moved to their own folder
