+++ voltixApp/VoltixApp/Chains/common.swift
@@ -74,23 +74,3 @@ struct SignatureProvider {
         return Data()
     }
 }
-
-extension Int64 {
-    func hexString() -> String {
-        var hexStr = String(format: "%02x", self)
-        if hexStr.count % 2 != 0 {
-            hexStr = "0" + hexStr
-        }
-        return hexStr
-    }
-}
-
-extension Data{
-    func sha256() -> Data {
-        var hash = [UInt8](repeating: 0,  count: Int(CC_SHA256_DIGEST_LENGTH))
-        self.withUnsafeBytes {
-            _ = CC_SHA256($0.baseAddress, CC_LONG(self.count), &hash)
-        }
-        return Data(hash)
-    }
-}
