+++ voltixApp/VoltixApp/Extensions/DataExtension.swift
@@ -1,0 +1,21 @@
+//
+//  DataExntension.swift
+//  VoltixApp
+//
+//  Created by Amol Kumar on 2024-03-04.
+//
+
+import OSLog
+import CommonCrypto
+
+extension Data{
+    func sha256() -> Data {
+        var hash = [UInt8](repeating: 0,  count: Int(CC_SHA256_DIGEST_LENGTH))
+        
+        self.withUnsafeBytes {
+            _ = CC_SHA256($0.baseAddress, CC_LONG(self.count), &hash)
+        }
+        
+        return Data(hash)
+    }
+}
