+++ voltixApp/VoltixApp/Model/Vault.swift
@@ -94,21 +94,3 @@ struct KeyShare: Codable {
     let pubkey: String
     let keyshare: String
 }
-
-// define some functions used for test
-extension Vault {
-    static func loadTestData(modelContext: ModelContext) {
-        modelContext.insert(Vault(name: "test", signers: ["A", "B", "C"], pubKeyECDSA: "ECDSA PubKey", pubKeyEdDSA: "EdDSA PubKey", keyshares: [KeyShare](), localPartyID: "first", hexChainCode: ""))
-        modelContext.insert(Vault(name: "test 1", signers: ["C", "D", "E"], pubKeyECDSA: "ECDSA PubKey", pubKeyEdDSA: "EdDSA PubKey", keyshares: [KeyShare](), localPartyID: "second", hexChainCode: ""))
-    }
-    
-    static var sampleVaults: () throws -> ModelContainer = {
-        let schema = Schema([Vault.self])
-        let configuration = ModelConfiguration(isStoredInMemoryOnly: true)
-        let container = try ModelContainer(for: schema, configurations: [configuration])
-        Task { @MainActor in
-            Vault.loadTestData(modelContext: container.mainContext)
-        }
-        return container
-    }
-}
