+++ voltixApp/VoltixApp/Services/Web3Service.swift
@@ -189,23 +189,3 @@ class Web3Service: ObservableObject {
     }
     
 }
-
-    // MARK: - String Extensions for Padding and Hex Processing
-
-extension String {
-    func paddingLeft(toLength: Int, withPad character: String) -> String {
-        let toPad = toLength - self.count
-        if toPad < 1 {
-            return self
-        }
-        
-        return "".padding(toLength: toPad, withPad: character, startingAt: 0) + self
-    }
-    
-    func stripHexPrefix() -> String {
-        if self.hasPrefix("0x") {
-            return String(self.dropFirst(2))
-        }
-        return self
-    }
-}
