+++ voltixApp/VoltixApp/Utils/Utils.swift
@@ -152,29 +152,3 @@ enum Utils {
         return UIDevice.current.name
     }
 }
-
-
-extension BigUInt {
-        /// Initializes a new BigUInt from an Int.
-        /// - Parameter value: The Int value to convert. Must be non-negative.
-    public init(_ value: Int) {
-        precondition(value >= 0, "BigUInt cannot represent negative values")
-        self.init(words: [Word(value)])
-    }
-}
-extension BigUInt {
-        /// Initializes a new BigUInt from a String.
-        /// - Parameter string: The String representation of the number.
-        /// - Parameter radix: The base of the number in the string. Default is 10.
-    public init?(_ string: String, radix: Int = 10) {
-        guard !string.isEmpty else { return nil }
-        self.init()
-        
-        var bigInt = BigUInt()
-        for character in string {
-            guard let digit = Int(String(character), radix: radix) else { return nil }
-            bigInt = bigInt.multiplied(by: BigUInt(radix)) + BigUInt(digit)
-        }
-        self = bigInt
-    }
-}
